requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
supabase==2.3.0
openai==1.12.0
schedule==1.2.0
logging==0.4.9.6
boto3==1.34.11
aws-lambda-powertools==2.32.0 